+++ code/modules/detectivework/tools/sample_kits.dm
@@ -1,6 +1,7 @@
 /obj/item/weapon/sample
 	name = "forensic sample"
 	icon = 'icons/obj/forensics.dmi'
+	item_flags = ITEM_FLAG_NO_PRINT
 	w_class = ITEM_SIZE_TINY
 	var/list/evidence = list()
 
@@ -24,7 +25,7 @@
 	if(!supplied.evidence || !supplied.evidence.len)
 		return 0
 	evidence |= supplied.evidence
-	name = "[initial(name)] (combined)"
+	SetName("[initial(name)] (combined)")
 	to_chat(user, "<span class='notice'>You transfer the contents of \the [supplied] into \the [src].</span>")
 	return 1
 
@@ -36,7 +37,7 @@
 			evidence[print] = stringmerge(evidence[print],supplied.evidence[print])
 		else
 			evidence[print] = supplied.evidence[print]
-	name = "[initial(name)] (combined)"
+	SetName("[initial(name)] (combined)")
 	to_chat(user, "<span class='notice'>You overlay \the [src] and \the [supplied], combining the print records.</span>")
 	return 1
 
@@ -46,8 +47,7 @@
 
 /obj/item/weapon/sample/attackby(var/obj/O, var/mob/user)
 	if(O.type == src.type)
-		user.unEquip(O)
-		if(merge_evidence(O, user))
+		if(user.unEquip(O) && merge_evidence(O, user))
 			qdel(O)
 		return 1
 	return ..()
@@ -77,7 +77,7 @@
 	to_chat(user, "<span class='notice'>You firmly press your fingertips onto the card.</span>")
 	var/fullprint = H.get_full_print()
 	evidence[fullprint] = fullprint
-	name = "[initial(name)] (\the [H])"
+	SetName("[initial(name)] (\the [H])")
 	icon_state = "fingerprint1"
 
 /obj/item/weapon/sample/print/attack(var/mob/living/M, var/mob/user)
@@ -114,7 +114,7 @@
 		var/fullprint = H.get_full_print()
 		evidence[fullprint] = fullprint
 		copy_evidence(src)
-		name = "[initial(name)] (\the [H])"
+		SetName("[initial(name)] (\the [H])")
 		icon_state = "fingerprint1"
 		return 1
 	return 0
@@ -124,6 +124,9 @@
 		for(var/print in supplied.fingerprints)
 			evidence[print] = supplied.fingerprints[print]
 		supplied.fingerprints.Cut()
+
+/obj/item/weapon/forensics
+	item_flags = ITEM_FLAG_NO_PRINT
 
 /obj/item/weapon/forensics/sample_kit
 	name = "fiber collection kit"
@@ -143,7 +146,7 @@
 /obj/item/weapon/forensics/sample_kit/afterattack(var/atom/A, var/mob/user, var/proximity)
 	if(!proximity)
 		return
-	if(can_take_sample(user, A))
+	if(user.skill_check(SKILL_FORENSICS, SKILL_ADEPT) && can_take_sample(user, A))
 		take_sample(user,A)
 		. = 1
 	else
